+++ SmartPharmacySystemBackend/SmartPharmacySystem.Application/DTOs/StockCounts/CreateStockCountItemDto.cs
@@ -4,7 +4,7 @@ public class CreateStockCountItemDto
 {
     public int MedicineId { get; set; }
     public string BatchNumber { get; set; } = string.Empty;
-    public int SystemQuantity { get; set; }
-    public int CountedQuantity { get; set; }
+    public decimal SystemQuantity { get; set; }
+    public decimal CountedQuantity { get; set; }
     public string? Notes { get; set; }
 }
