+++ SmartPharmacySystemBackend/SmartPharmacySystem.Application/DTOs/StockMovement/StockCardDto.cs
@@ -6,8 +6,8 @@ public class StockCardDto
 {
     public DateTime Date { get; set; }
     public StockMovementType MovementType { get; set; }
-    public int QuantityChange { get; set; }
+    public decimal QuantityChange { get; set; }
     public string ReferenceNumber { get; set; } = string.Empty;
-    public int RunningBalance { get; set; }
+    public decimal RunningBalance { get; set; }
     public string? Notes { get; set; }
 }
